session_3

+++ session2/session2_8/main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-unsigned long fact(int);
+unsigned int fact(int);
 int main() {
     int number;
     cout << "enter a positive integer: ";
@@ -11,7 +11,8 @@ int main() {
     return 0;
 }
 
-unsigned long fact(int x){
+unsigned int fact(int x){
     if (x != 0) return (x * fact(x - 1));
+//    5 * fact(4) ---> 5 * 4 * fact(3) ---> 5 * 4 * 3 * fact(2) ---> 5 * 4 * 3 * 2 * fact(1) --->5 * 4 * 3 * 2 * 1 = 120
     return 1;
 }
